+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # druxt-auth
 
+## 0.2.0
+
+### Minor Changes
+
+- 3ef518e: feat(#8): added support for password grant.
+
+### Patch Changes
+
+- 5bab97a: Enabled the Nuxt Vuex store.
+
 ## 0.1.0
 
 ### Minor Changes